@@ -6,18 +6,19 @@
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@react-three/drei": "^7.3.1",
-    "@react-three/fiber": "^7.0.6",
-    "@reduxjs/toolkit": "^1.7.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "@react-three/drei": "^9.16.0",
+    "@react-three/fiber": "^8.0.0",
+    "@reduxjs/toolkit": "^1.8.3",
+    "react": "^18.0",
+    "react-dom": "^18.0",
     "react-redux": "^7.2.6",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "redux-undo": "^1.0.1",
-    "three": "^0.131.3",
+    "three": "^0.133",
+    "three-stdlib": "^2.12.1",
     "txml": "^5.1.1",
     "workbox-cacheable-response": "^6.4.1",
     "workbox-expiration": "^6.4.1",
